amazed with the rumor of  Aso sickened.
this is rumor.
Aso and parliamentarians sent money to Clinton conglomerates with political contributions as  Akishinomiya will do emperor.  because the child of Princess Diana is convenient. Britain also has only the eldest son to bear the name of the royal family pretty strict with precepts. If I say it in Japan it is the prince of the eldest son. i don't understand Mr. Aso's plan. there is no difference in that Hironomiya is the eldest son of Michiko.  Aso is been so worring by hironomiya of International Convention decided. because Aso would considered that Japanese could goes to the north south?  
However, if you we force attack north, ballistic missiles will fly and abductee victims will have dead people. · We are meaningless north attack that we hope only to back to Japan all abductee victims.
it is been not a joke with Aso's rumor.